backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
supabase==2.4.0
httpx==0.24.1
slowapi==0.1.9
scikit-learn==1.4.2
python-dotenv==1.0.1